requirements.txt
asgiref==3.5.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.1
django-appconf==1.0.6
django-compressor==4.3.1
django-model-utils==4.3.1
django-phonenumber-field
django-phonenumbers==1.0.1
djangorestframework==3.14.0
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.7
Pillow==9.4.0
platformdirs==3.1.1
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.6
rcssmin==1.1.1
requests==2.28.2
rjsmin==1.2.1
sqlparse==0.4.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.14
whitenoise==6.4.0
zipp==3.15.0
psycopg2-binary==2.9.9
drf_yasg==1.21.7
django-celery-beat==2.5.0
django-celery-results==2.4.0
celery==5.2.7
flower==2.0
redis==4.3.4
django-cors-headers